feat: 주문 확정 시 고객 통계 자동 업데이트 시스템 구현   - Order 도메인에 savedMoney, savedCo2 필드 추가   - OrderSavingsCalculationService 구현 (절약 금액/CO2 계산)   - CustomerStatsUpdateService 구현 (고객 통계 업데이트)   - SavingsResult VO를 domain/model/vo 패키지로 이동   - Money 타입을 Integer에서 Long으로 변경 (대용량 금액 처리)   - 주문 생성 시 절약 값 사전 계산, 확정 시 고객 통계 업데이트   - 기존 CustomerStatsCalculator의 중복 메서드 제거

+++ src/main/java/com/kkulddip/order/presentation/rest/dto/response/CreateOrderResponse.java
@@ -9,8 +9,8 @@ public record CreateOrderResponse(
     String orderId,      // Long -> String으로 변경
     Long customerId,
     Long storeId,
-    Integer originalPrice,
-    Integer finalPrice,
+    Long originalPrice,
+    Long finalPrice,
     String orderStatus,
     LocalDateTime orderDate
 ) {}
